backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Anthropic Claude SDK
anthropic>=0.40.0

# Langfuse (LLM Observability)
langfuse==2.58.3

# Graphiti (for knowledge graph via MCP)
graphiti-core==0.3.8

# Neo4j driver
neo4j==5.26.0

# Document processing
llama-parse==0.5.16  # Intelligent PDF parsing for complex documents
PyPDF2==3.0.1  # Fallback PDF parser
python-docx==1.1.2
pypandoc==1.14

# OpenAI (for Graphiti embeddings)
openai==1.55.3

# Utilities
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
aiofiles==24.1.0

# CORS
fastapi-cors==0.0.6
